Logbooks components added
