+++ frontend/src/components/Logbooks/AddLogbookCard.css
@@ -1,0 +1,37 @@
+.add-logbook-card {
+    border-radius: 30px;
+    max-width: 300px;
+    padding: 20px;
+    position: relative;
+    cursor: pointer;
+    transition: transform 0.2s;
+    height: 100%;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    gap: 8px;
+  }
+  
+  .add-logbook-card:hover {
+    transform: translateY(-4px);
+  }
+  
+  /* Text Styles */
+  .add-logbook-card span {
+    color: #1f2937;
+    font-size: 14px;
+    font-weight: 400;
+  }
+  
+  /* Icon Styles */
+  .plus-circle-icon {
+    width: 40px;
+    height: 40px;
+    color: #244b94;
+    padding: 8px;
+    border: 4px solid #244b94;
+    border-radius: 50%;
+    stroke-width: 3px;
+  }
+
