+++ frontend/src/components/Logbooks/LogbookCard.css
@@ -1,0 +1,134 @@
+/* Base Card Styles */
+.logbook-card {
+    border-radius: 30px;
+    max-width: 300px;
+    padding: 20px;
+    position: relative;
+    cursor: pointer;
+    transition: transform 0.2s;
+    background: #ebf2fd;
+    backdrop-filter: blur(20px);
+    box-shadow:
+      0px 16px 60px 0px #cdd8f040,
+      0px 0px 60px 0px #ffffff80 inset;
+    height: 300px;
+  }
+  
+  .logbook-card:hover .book-cover {
+    transform: translateY(-10px);
+  }
+  
+  /* Logbook Card Variants */
+  .logbook-card.congenital {
+    background: #fbf4ff;
+  }
+  
+  .logbook-card.ophthalmology {
+    background: #ebfbfe;
+  }
+  
+  .logbook-card.obstetrics {
+    background: #f0edf3;
+  }
+  
+  .logbook-card.general-surgery {
+    background: #e8fdff;
+  }
+  
+  /* Book Cover Styles */
+  .book-cover {
+    position: relative;
+    height: 100%;
+    overflow: hidden;
+    transition: transform 0.2s;
+  }
+  
+  .book-cover-image {
+    width: 60%;
+    margin-top: 10%;
+    margin-left: 20%;
+    object-fit: cover;
+    clip-path: inset(0 0 30% 0);
+  }
+  
+  .book-cover-image.obstetrics {
+    width: 30%;
+    margin-top: 90%;
+  }
+  
+  .log-rectangle {
+    width: 170%;
+    height: 240%;
+    padding: 13px;
+    object-fit: cover;
+    position: absolute;
+    bottom: 0;
+    right: -57%;
+  }
+  
+  /* Details Container */
+  .details-container {
+    position: relative;
+    height: 140px;
+  }
+  
+  .book-details {
+    position: relative;
+    padding: 10px;
+    bottom: 100%;
+    z-index: 2;
+    align-items: flex-start;
+  }
+  
+  /* Text Styles */
+  .book-title {
+    font-size: 22px;
+    font-weight: bold;
+    color: #1e1e1e;
+    margin-bottom: 12px;
+    text-align: left;
+  }
+  
+  .type-label,
+  .storage-info {
+    font-size: 14px;
+    font-weight: 500;
+    text-align: left;
+    display: flex;
+    align-items: center;
+    gap: 4px;
+  }
+  
+  .type-label {
+    color: #333333;
+  }
+  
+  .storage-info {
+    margin: 8px 0;
+  }
+  
+  .type-value,
+  .storage-count {
+    font-size: 14px;
+    font-weight: 500;
+  }
+  
+  .type-value {
+    color: #244b94;
+  }
+  
+  .storage-count {
+    color: #5d9eff;
+  }
+  
+  .created-date {
+    font-size: 10px;
+    color: #656b6f;
+    margin-top: 8px;
+    text-align: left;
+  }
+  
+  .created-date strong {
+    font-weight: 600;
+  }
+
